+++ compiler/src/vm/heap/mod.rs
@@ -125,7 +125,6 @@ impl Heap {
         &self,
         other: &mut Heap,
         addresses: &[Pointer],
-        channel_map: &HashMap<ChannelId, ChannelId>,
     ) -> Vec<Pointer> {
         let mut objects_to_refcounts = HashMap::new();
         for address in addresses {
@@ -147,7 +146,7 @@ impl Heap {
                 address_map[&address],
                 Object {
                     reference_count: refcount,
-                    data: Self::map_data(&address_map, channel_map, &self.get(address).data),
+                    data: Self::map_data(&address_map, &self.get(address).data),
                 },
             );
         }
@@ -168,11 +167,7 @@ impl Heap {
             self.gather_objects_to_clone(objects_to_refcounts, child);
         }
     }
-    fn map_data(
-        address_map: &HashMap<Pointer, Pointer>,
-        channel_map: &HashMap<ChannelId, ChannelId>,
-        data: &Data,
-    ) -> Data {
+    fn map_data(address_map: &HashMap<Pointer, Pointer>, data: &Data) -> Data {
         match data {
             Data::Int(int) => Data::Int(int.clone()),
             Data::Text(text) => Data::Text(text.clone()),
@@ -194,19 +189,12 @@ impl Heap {
                 body: closure.body.clone(),
             }),
             Data::Builtin(builtin) => Data::Builtin(builtin.clone()),
-            Data::SendPort(port) => Data::SendPort(SendPort::new(channel_map[&port.channel])),
-            Data::ReceivePort(port) => {
-                Data::ReceivePort(ReceivePort::new(channel_map[&port.channel]))
-            }
+            Data::SendPort(port) => Data::SendPort(SendPort::new(port.channel)),
+            Data::ReceivePort(port) => Data::ReceivePort(ReceivePort::new(port.channel)),
         }
     }
-    pub fn clone_single_to_other_heap(
-        &self,
-        other: &mut Heap,
-        address: Pointer,
-        channel_map: &HashMap<ChannelId, ChannelId>,
-    ) -> Pointer {
-        self.clone_multiple_to_other_heap(other, &[address], channel_map)
+    pub fn clone_single_to_other_heap(&self, other: &mut Heap, address: Pointer) -> Pointer {
+        self.clone_multiple_to_other_heap(other, &[address])
             .pop()
             .unwrap()
     }
