+++ compiler/src/vm/mod.rs
@@ -1,9 +1,11 @@
 mod builtin_functions;
 mod channel;
+mod fiber;
 mod heap;
 pub mod tracer;
 pub mod use_provider;
-mod vm;
+pub mod vm;
 
+pub use fiber::{Fiber, TearDownResult};
 pub use heap::{Closure, Heap, Object, Pointer};
-pub use vm::{Status, TearDownResult, Vm};
+pub use vm::Vm;
